test: promote connection_config_test.rb into a standalone unit test class
Reviewer feedback: this suite stubs the FFI connect call and never touches
a server, so it belongs in test/unit alongside Utils and
OpenTelemetryProvider rather than mixed into the standalone/cluster
integration test classes.
