+++ test/integration/cluster/cluster_commands_test.rb
@@ -31,7 +31,6 @@ class TestClusterCommands < Minitest::Test
 
   # ValkeyTests modules without conflicting setup methods
   include ValkeyTests::Bitpos
-  include ValkeyTests::ConnectionConfig
   include ValkeyTests::Call
   include ValkeyTests::GenericCommands
   include ValkeyTests::Scanning
