+++ test/integration/standalone/valkey_test.rb
@@ -10,7 +10,6 @@ class TestStandaloneValkey < Minitest::Test
   # ValkeyTests modules for valkey-glide-ruby specific functionality
   include ValkeyTests::AuthCommands
   include ValkeyTests::Bitpos
-  include ValkeyTests::ConnectionConfig
   include ValkeyTests::Call
   include ValkeyTests::ConnectionLifecycle
   include ValkeyTests::FailoverCommands
